Initial Commit
Initial commit is somewhat of a lie. This is a reset of all the commits
in this repository. This is primarily because commit messages like "More
changes" are not useful.

This resets all those messages away. Hopefully the code itself is
self-explanatory, or comments are up-to-date and explain the code well.

For the unlikely chance that someone has a repo that is based off this,
I'm, sorry, but hopefully the future of these configuration repositories
will have more useful commit messages that convey the reasoning behind
changes.